AST working???

+++ src/backend/NumberExpr.hpp
@@ -7,11 +7,11 @@
 
 class NumberExpr : public ASTNode {
 public:
-    int value;
+    double value;
 
-    NumberExpr(int v) : value(v) {}
+    NumberExpr(double v) : value(v) {}
 
-    void print(int indent = 0) const override {
+    void print(int indent = 0) const override {  // <-- corrigido de double para int
         std::cout << std::string(indent, ' ') << "Number(" << value << ")\n";
     }
 };
